test: add tests/pg_cron suite (app installed, pg_cron db); move pg_cron tests + fixtures
