+++ tests/pg_cron/test_pg_cron_post_migrate.py
@@ -12,7 +12,6 @@ from django_absurd.queues import get_absurd_client
 
 pytestmark = [
     pytest.mark.django_db(transaction=True),
-    pytest.mark.pg_cron,
     pytest.mark.usefixtures("ensure_pg_cron", "_clear_owned_pg_cron_jobs"),
 ]
