+++ tests/pg_cron/test_pg_cron_sync_jobs.py
@@ -14,7 +14,6 @@ from django_absurd.pg_cron.reconcile import (
 
 pytestmark = [
     pytest.mark.django_db(transaction=True),
-    pytest.mark.pg_cron,
     pytest.mark.usefixtures("ensure_pg_cron", "_clear_owned_pg_cron_jobs"),
 ]
