+++ tests/pg_cron/test_pg_cron_sync_rows.py
@@ -4,7 +4,7 @@ from django_absurd.backends import get_absurd_backends
 from django_absurd.pg_cron.models import ScheduledTask
 from django_absurd.pg_cron.reconcile import sync_crons
 
-pytestmark = [pytest.mark.django_db(transaction=True), pytest.mark.pg_cron]
+pytestmark = pytest.mark.django_db(transaction=True)
 
 ABSURD = "django_absurd.backends.AbsurdBackend"
